Fix(cpp00/ex00) : int -> size_t

+++ cpp00/ex00/Megaphone.cpp
@@ -14,7 +14,7 @@ int main(int argc, char **argv)
     for (int i = 1; i < argc; ++i)
     {
         std::string str = argv[i];
-        for (int j = 0; j < str.length(); ++j)
+        for (size_t j = 0; j < str.length(); ++j)
         {
             str[j] = std::toupper(str[j]);
         }
